videoscale: Change navigation events to scale and respect internal region without borders
The current version of videoscale scales navigation events based on the input to output ratio.
If the output includes borders, the scaling includes the border area, which causes navigation
events to be misaligned compared to the position of the mouse. This patch provides a fix for
this behavior, scaling the navigation events based on only the internal region and not the
borders. All events that are within the borders are discarded.

https://bugzilla.gnome.org/show_bug.cgi?id=762986